backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.35.63
python-jose[cryptography]==3.3.0
pydantic==2.10.2
pydantic-settings==2.6.1
